requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1

# RAG stack
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
transformers>=4.42.0
torch>=2.2.0

# Utilities
numpy>=1.26.4
pandas>=2.2.2

# UI
streamlit>=1.36.0

# Testing
pytest>=8.0.0
